+++ modules/@angular/common/test/directives/ng_style_spec.ts
@@ -7,7 +7,7 @@
  */
 
 import {beforeEach, beforeEachProviders, ddescribe, xdescribe, describe, expect, iit, inject, it, xit,} from '@angular/core/testing/testing_internal';
-import {TestComponentBuilder, ComponentFixture} from '@angular/compiler/testing';
+import {TestComponentBuilder} from '@angular/compiler/testing';
 import {AsyncTestCompleter} from '@angular/core/testing/testing_internal';
 
 import {StringMapWrapper} from '../../src/facade/collection';
