+++ modules/@angular/common/test/directives/ng_template_outlet_spec.ts
@@ -7,7 +7,7 @@
  */
 
 import {beforeEach, ddescribe, describe, expect, iit, inject, it, xit,} from '@angular/core/testing/testing_internal';
-import {TestComponentBuilder, ComponentFixture} from '@angular/compiler/testing';
+import {TestComponentBuilder} from '@angular/compiler/testing';
 import {AsyncTestCompleter} from '@angular/core/testing/testing_internal';
 import {Component, Directive, TemplateRef, ContentChildren, QueryList} from '@angular/core';
 import {NgTemplateOutlet} from '@angular/common';
