+++ modules/@angular/core/test/linker/view_injector_integration_spec.ts
@@ -7,8 +7,8 @@
  */
 
 import {describe, ddescribe, it, iit, xit, xdescribe, expect, beforeEach, beforeEachProviders, inject,} from '@angular/core/testing/testing_internal';
-import {fakeAsync, flushMicrotasks, tick} from '@angular/core/testing';
-import {TestComponentBuilder, ComponentFixture} from '@angular/compiler/testing';
+import {fakeAsync, flushMicrotasks, tick, ComponentFixture} from '@angular/core/testing';
+import {TestComponentBuilder} from '@angular/compiler/testing';
 import {isBlank} from '../../src/facade/lang';
 import {Type, ViewContainerRef, TemplateRef, ElementRef, ChangeDetectorRef, ChangeDetectionStrategy, Directive, Component, DebugElement, forwardRef, Input, PipeTransform, Attribute, ViewMetadata, provide, Optional, Inject, Self, InjectMetadata, Pipe, Host, SkipSelfMetadata} from '@angular/core';
 import {NgIf, NgFor} from '@angular/common';
