+++ modules/@angular/core/testing.ts
@@ -7,6 +7,7 @@
  */
 
 export * from './testing/async';
+export * from './testing/component_fixture';
 export * from './testing/fake_async';
 export * from './testing/test_injector';
 export * from './testing/testing';
