+++ tools/public_api_guard/core/testing.d.ts
@@ -6,6 +6,23 @@ export declare function beforeEach(fn: Function): void;
 
 export declare function beforeEachProviders(fn: () => Array<any>): void;
 
+export declare class ComponentFixture<T> {
+    changeDetectorRef: ChangeDetectorRef;
+    componentInstance: any;
+    componentRef: ComponentRef<T>;
+    debugElement: DebugElement;
+    elementRef: ElementRef;
+    nativeElement: any;
+    ngZone: NgZone;
+    constructor(componentRef: ComponentRef<T>, ngZone: NgZone, autoDetect: boolean);
+    autoDetectChanges(autoDetect?: boolean): void;
+    checkNoChanges(): void;
+    destroy(): void;
+    detectChanges(checkNoChanges?: boolean): void;
+    isStable(): boolean;
+    whenStable(): Promise<any>;
+}
+
 export declare var ddescribe: Function;
 
 export declare var describe: Function;
